Cafe Level Design
added second floor & stairs. textured almost everything at this point, save for the walls. (maybe pure white is the best choice? artistic/stylistic decision to be made) added barricade to 2nd floor staircase (possibly block of seconf floor as a means to difficulty scale the game? (more seats to attend to later on)